move db out of api directory
